Rename serializer to match folder structure better
